+++ Android/core/data/src/main/java/com/core/data/di/DataModule.kt
@@ -3,9 +3,9 @@ package com.core.data.di
 import com.core.data.repository.auth.LoginRepositoryImpl
 import com.core.data.repository.auth.SignupRepositoryImpl
 import com.core.data.repository.bluetooth.BluetoothRepositoryImpl
+import com.core.data.repository.network.GetHeartRateRepositoryImpl
 import com.core.data.repository.prefs.AuthPrefsRepositoryImpl
 import com.core.data.repository.prefs.DevicePrefsRepositoryImpl
-import com.core.data.repository.statistics.GetHeartRateRepositoryImpl
 import com.core.data.repository.wearable.WearableRepositoryImpl
 import com.core.domain.repository.AuthPrefsRepository
 import com.core.domain.repository.BluetoothRepository
