+++ Android/core/data/src/main/java/com/core/data/repository/network/GetHeartRateRepositoryImpl.kt
@@ -1,11 +1,14 @@
-package com.core.data.repository.statistics
+package com.core.data.repository.network
 
+import com.core.data.mapper.toHeartRate
 import com.core.data.mapper.toHeartRateList
 import com.core.domain.repository.GetHeartRateRepository
 import com.core.firebase.HeartRateClient
 import com.core.model.HeartRate
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
@@ -24,4 +27,3 @@ class GetHeartRateRepositoryImpl @Inject constructor(
             it.toHeartRateList()
         }.flowOn(Dispatchers.IO)
     }
-}
